+++ src/consume.ts
@@ -20,6 +20,7 @@ import {
 } from "./clients/sidecar";
 import { registerCommandWithLogging } from "./commands";
 import { Logger } from "./logging";
+import { getTelemetryLogger } from "./telemetry";
 import { getSidecar, type SidecarHandle } from "./sidecar";
 import { BitSet, Stream, includesSubstring } from "./stream/stream";
 import { handleWebviewMessage } from "./webview/comms/comms";
@@ -124,7 +125,7 @@ function messageViewerStartPollingCommand(
   /** Most recent response payload from Consume API. */
   const latestResult = os.signal<SimpleConsumeMultiPartitionResponse | null>(null);
   /** Most recent failure info */
-  const latestError = os.signal<string[] | null>(null);
+  const latestError = os.signal<{ message: string } | null>(null);
   /** Timestamp of the most recent successful consumption request. */
   const latestFetch = os.signal<number>(0);
 
@@ -323,10 +324,11 @@ function messageViewerStartPollingCommand(
           flushMessages(stream);
           latestResult(result);
           latestFetch(Date.now());
+          latestError(null);
           notifyUI();
         });
       } catch (error) {
-        let reportable: any = null;
+        let reportable: { message: string } | null = null;
         let shouldPause = false;
         /* Async operations can be aborted by provided AbortController that is
         controlled by the watcher. Nothing to log in this case. */
@@ -337,15 +339,42 @@ function messageViewerStartPollingCommand(
           const payload = await error.response.json();
           // FIXME: this response error coming from the middleware that has to be present to avoid openapi error about missing middlewares
           if (!payload?.aborted) {
-            shouldPause = error.response.status >= 400 && error.response.status < 500;
-            reportable = JSON.stringify(payload);
+            const status = error.response.status;
+            shouldPause = status >= 400;
+            switch (status) {
+              case 401: {
+                reportable = { message: "Authentication required." };
+                break;
+              }
+              case 403: {
+                reportable = { message: "Insufficient permissions to read from topic." };
+                break;
+              }
+              case 404: {
+                if (String(payload?.title).startsWith("Error fetching the messages")) {
+                  reportable = { message: "Topic not found." };
+                } else {
+                  reportable = { message: "Unable to connect to the server." };
+                }
+                break;
+              }
+              case 429: {
+                reportable = { message: "Too many requests. Try again later." };
+                break;
+              }
+              default: {
+                reportable = { message: "Something went wrong." };
+                break;
+              }
+            }
             logger.error(
               `An error occurred during messages consumption. Status ${error.response.status}`,
             );
           }
         } else if (error instanceof Error) {
           logger.error(error.message);
-          reportable = error.message;
+          reportable = { message: "An internal error occurred." };
+          shouldPause = true;
         }
 
         os.batch(() => {
@@ -356,9 +385,7 @@ function messageViewerStartPollingCommand(
             timer((timer) => timer.pause());
           }
           if (reportable != null) {
-            latestError((errors) => {
-              return errors == null ? [reportable] : [reportable].concat(errors).slice(0, 10);
-            });
+            latestError(reportable);
           }
           notifyUI();
         });
@@ -434,6 +461,7 @@ function messageViewerStartPollingCommand(
         return (search?.query ?? "") satisfies MessageResponse<"GetSearchQuery">;
       }
       case "PreviewMessageByIndex": {
+        track({ action: "preview-message" });
         const { messages, serialized } = stream();
         const index = body.index;
         const message = messages.at(index);
@@ -459,6 +487,7 @@ function messageViewerStartPollingCommand(
         return null;
       }
       case "PreviewJSON": {
+        track({ action: "preview-snapshot" });
         const {
           timestamp,
           messages: { values },
@@ -487,6 +516,7 @@ function messageViewerStartPollingCommand(
         return null satisfies MessageResponse<"PreviewJSON">;
       }
       case "SearchMessages": {
+        track({ action: "search" });
         if (body.search != null) {
           const { capacity, messages } = stream();
           const values = messages.values;
@@ -524,6 +554,7 @@ function messageViewerStartPollingCommand(
         return null satisfies MessageResponse<"StreamResume">;
       }
       case "ConsumeModeChange": {
+        track({ action: "consume-mode-change" });
         mode(body.mode);
         const maxPollRecords = Math.min(DEFAULT_MAX_POLL_RECORDS, os.peek(stream).capacity);
         params(getParams(body.mode, body.timestamp, maxPollRecords));
@@ -542,6 +573,7 @@ function messageViewerStartPollingCommand(
         return null satisfies MessageResponse<"ConsumeModeChange">;
       }
       case "PartitionConsumeChange": {
+        track({ action: "consume-partition-change" });
         partitionConsumed(body.partitions);
         const maxPollRecords = Math.min(DEFAULT_MAX_POLL_RECORDS, os.peek(stream).capacity);
         params((value) => getParams(os.peek(mode), value.timestamp, maxPollRecords));
@@ -560,16 +592,19 @@ function messageViewerStartPollingCommand(
         return null satisfies MessageResponse<"PartitionConsumeChange">;
       }
       case "PartitionFilterChange": {
+        track({ action: "filter-partition-change" });
         partitionFilter(body.partitions);
         notifyUI();
         return null satisfies MessageResponse<"PartitionFilterChange">;
       }
       case "TimestampFilterChange": {
+        debouncedTrack({ action: "filter-timestamp-change" });
         timestampFilter(body.timestamps);
         notifyUI();
         return null satisfies MessageResponse<"TimestampFilterChange">;
       }
       case "MessageLimitChange": {
+        track({ action: "consume-message-limit-change" });
         const maxPollRecords = Math.min(DEFAULT_MAX_POLL_RECORDS, body.limit);
         params((value) => getParams(os.peek(mode), value.timestamp, maxPollRecords));
         stream(new Stream(body.limit));
@@ -711,4 +746,14 @@ class Timer extends Data {
   reset(this: Timer) {
     return this.copy({ start: Date.now(), offset: 0 });
   }
+}
+
+function track(details: object) {
+  getTelemetryLogger().logUsage("Message Viewer Action", details);
+}
+
+let timer: ReturnType<typeof setTimeout>;
+function debouncedTrack(details: object) {
+  clearTimeout(timer);
+  timer = setTimeout(track, 200, details);
 }